+++ src/snowflake/snowpark/mock/_plan.py
@@ -39,6 +39,7 @@ from snowflake.snowpark._internal.analyzer.window_expression import (
     UnboundedPreceding,
     WindowExpression,
 )
+from snowflake.snowpark.mock._udf_utils import types_are_compatible
 from snowflake.snowpark.mock._util import get_fully_qualified_name
 from snowflake.snowpark.mock._window_utils import (
     EntireWindowIndexer,
@@ -52,7 +53,6 @@ if TYPE_CHECKING:
 
 from contextlib import ExitStack
 
-from snowflake.connector.options import pandas as pd
 from snowflake.snowpark._internal.analyzer.analyzer_utils import (
     EXCEPT,
     INTERSECT,
@@ -137,6 +137,7 @@ from snowflake.snowpark._internal.utils import (
 )
 from snowflake.snowpark.column import Column
 from snowflake.snowpark.mock._functions import _MOCK_FUNCTION_IMPLEMENTATION_MAP
+from snowflake.snowpark.mock._options import pandas as pd
 from snowflake.snowpark.mock._select_statement import (
     MockSelectable,
     MockSelectableEntity,
@@ -294,7 +295,7 @@ def handle_range_frame_indexing(
     res_index: "pd.Index",
     res: "pd.api.typing.DataFrameGroupBy",
     analyzer: "MockAnalyzer",
-    expr_to_alias: Optional[Dict[str, str]],
+    expr_to_alias: Dict[str, str],
     unbounded_preceding: bool,
     unbounded_following: bool,
 ) -> "pd.api.typing.RollingGroupby":
@@ -338,7 +339,7 @@ def handle_function_expression(
     exp: FunctionExpression,
     input_data: Union[TableEmulator, ColumnEmulator],
     analyzer: "MockAnalyzer",
-    expr_to_alias: Optional[Dict[str, str]],
+    expr_to_alias: Dict[str, str],
     current_row=None,
 ):
     # Special case for count_distinct
@@ -360,6 +361,11 @@ def handle_function_expression(
             exp.udf_name = udf_name
             return handle_udf_expression(
                 exp, input_data, analyzer, expr_to_alias, current_row
+            )
+
+        if exp.api_call_source == "functions.call_udf":
+            raise SnowparkLocalTestingException(
+                f"Unknown function {func_name}. UDF by that name does not exist."
             )
 
         # this is missing function in snowpark-python, need support for both live and local test
@@ -384,18 +390,33 @@ def handle_function_expression(
         type_hint = str(type_hints[key])
         keep_literal = "Column" not in type_hint
         if key == spec.varargs:
-            to_pass_args.extend(
-                [
-                    calculate_expression(
-                        c,
-                        input_data,
-                        analyzer,
-                        expr_to_alias,
-                        keep_literal=keep_literal,
+            # SNOW-1441602: Move Star logic to calculate_expression once it can handle table returns
+            args = []
+            for c in exp.children[idx:]:
+                if isinstance(c, Star):
+                    args.extend(
+                        [
+                            calculate_expression(
+                                cc,
+                                input_data,
+                                analyzer,
+                                expr_to_alias,
+                                keep_literal=keep_literal,
+                            )
+                            for cc in c.expressions
+                        ]
                     )
-                    for c in exp.children[idx:]
-                ]
-            )
+                else:
+                    args.append(
+                        calculate_expression(
+                            c,
+                            input_data,
+                            analyzer,
+                            expr_to_alias,
+                            keep_literal=keep_literal,
+                        )
+                    )
+            to_pass_args.extend(args)
         else:
             try:
                 to_pass_args.append(
@@ -424,7 +445,13 @@ def handle_function_expression(
         )  # the row's 0-base index in the window
         to_pass_args.append(row_idx)
 
-    result = _MOCK_FUNCTION_IMPLEMENTATION_MAP[func_name](*to_pass_args)
+    try:
+        result = _MOCK_FUNCTION_IMPLEMENTATION_MAP[func_name](*to_pass_args)
+    except Exception as err:
+        SnowparkLocalTestingException.raise_from_error(
+            err,
+            error_message=f"Error executing mocked function '{func_name}'. See error traceback for detailed information.",
+        )
 
     # If none of the args are column emulators and the function result only has one item
     # assume that the single value should be repeated instead of Null filled. This allows
@@ -447,7 +474,7 @@ def handle_udf_expression(
     exp: FunctionExpression,
     input_data: Union[TableEmulator, ColumnEmulator],
     analyzer: "MockAnalyzer",
-    expr_to_alias: Optional[Dict[str, str]],
+    expr_to_alias: Dict[str, str],
     current_row=None,
 ):
     udf_registry = analyzer.session.udf._registry
@@ -520,13 +547,15 @@ def handle_udf_expression(
                 else:
                     shutil.copy2(module_path, temporary_import_path.name)
 
+        udf = udf_registry[udf_name]
+
         # Resolve handler callable
-        if type(udf_registry[udf_name]) is tuple:
-            module_name, handler_name = udf_registry[udf_name]
+        if type(udf.func) is tuple:
+            module_name, handler_name = udf.func
             exec(f"from {module_name} import {handler_name}")
             udf_handler = eval(handler_name)
         else:
-            udf_handler = udf_registry[udf_name]
+            udf_handler = udf.func
 
         # Compute
         function_input = TableEmulator(index=input_data.index)
@@ -535,23 +564,42 @@ def handle_udf_expression(
             function_input[col_name] = calculate_expression(
                 child, input_data, analyzer, expr_to_alias
             )
+        # Validate function inputs
+        actual_types = [col.datatype for col in function_input.sf_types.values()]
+        if len(actual_types) != len(udf._input_types):
+            raise SnowparkLocalTestingException(
+                f"Expected {len(udf._input_types)} arguments, but received {len(actual_types)}"
+            )
 
-        # we do not use pd.apply here because pd.apply will auto infer dtype for the output column
-        # this will lead to NaN or None information loss, think about the following case of a udf definition:
-        #    def udf(x): return numpy.sqrt(x) if x is not None else None
-        # calling udf(-1) and udf(None), pd.apply will infer the column dtype to be int which returns NaT for both
-        # however, we want NaT for the former case and None for the latter case.
-        # using dtype object + function execution does not have the limitation
-        # In the future maybe we could call fix_drift_between_column_sf_type_and_dtype in methods like set_sf_type.
-        # And these code would look like:
-        # res=input.apply(...)
-        # res.set_sf_type(ColumnType(exp.datatype, exp.nullable))  # fixes the drift and removes NaT
-        res = ColumnEmulator(
-            data=[udf_handler(*row) for _, row in function_input.iterrows()],
-            sf_type=ColumnType(exp.datatype, exp.nullable),
-            name=quote_name(f"{exp.udf_name}({', '.join(input_data.columns)})".upper()),
-            dtype=object,
-        )
+        for expected, actual in zip(udf._input_types, actual_types):
+            if not types_are_compatible(expected, actual):
+                raise SnowparkLocalTestingException(
+                    f"UDF received input types {actual_types}, but expected inputs types of {udf._input_types}"
+                )
+
+        try:
+            # we do not use pd.apply here because pd.apply will auto infer dtype for the output column
+            # this will lead to NaN or None information loss, think about the following case of a udf definition:
+            #    def udf(x): return numpy.sqrt(x) if x is not None else None
+            # calling udf(-1) and udf(None), pd.apply will infer the column dtype to be int which returns NaT for both
+            # however, we want NaT for the former case and None for the latter case.
+            # using dtype object + function execution does not have the limitation
+            # In the future maybe we could call fix_drift_between_column_sf_type_and_dtype in methods like set_sf_type.
+            # And these code would look like:
+            # res=input.apply(...)
+            # res.set_sf_type(ColumnType(exp.datatype, exp.nullable))  # fixes the drift and removes NaT
+            res = ColumnEmulator(
+                data=[udf_handler(*row) for _, row in function_input.iterrows()],
+                sf_type=ColumnType(exp.datatype, exp.nullable),
+                name=quote_name(
+                    f"{exp.udf_name}({', '.join(input_data.columns)})".upper()
+                ),
+                dtype=object,
+            )
+        except Exception as err:
+            SnowparkLocalTestingException.raise_from_error(
+                err, error_message=f"Python Interpreter Error: {err}"
+            )
 
         return res
 
@@ -563,7 +611,8 @@ def execute_mock_plan(
     import numpy as np
 
     if expr_to_alias is None:
-        expr_to_alias = {}
+        expr_to_alias = plan.expr_to_alias
+
     if isinstance(plan, (MockExecutionPlan, SnowflakePlan)):
         source_plan = plan.source_plan
         analyzer = plan.session._analyzer
@@ -664,7 +713,7 @@ def execute_mock_plan(
 
         if where:
             condition = calculate_expression(where, result_df, analyzer, expr_to_alias)
-            result_df = result_df[condition]
+            result_df = result_df[condition.fillna(value=False)]
 
         if order_by:
             result_df = handle_order_by_clause(
@@ -748,7 +797,7 @@ def execute_mock_plan(
             return entity_registry.read_table(entity_name)
         elif entity_registry.is_existing_view(entity_name):
             execution_plan = entity_registry.get_review(entity_name)
-            res_df = execute_mock_plan(execution_plan)
+            res_df = execute_mock_plan(execution_plan, expr_to_alias)
             return res_df
         else:
             db_schme_table = parse_table_name(entity_name)
@@ -757,7 +806,7 @@ def execute_mock_plan(
                 f"Object '{table}' does not exist or not authorized."
             )
     if isinstance(source_plan, Aggregate):
-        child_rf = execute_mock_plan(source_plan.child)
+        child_rf = execute_mock_plan(source_plan.child, expr_to_alias)
         if (
             not source_plan.aggregate_expressions
             and not source_plan.grouping_expressions
@@ -1061,7 +1110,7 @@ def execute_mock_plan(
 
             condition = calculate_expression(
                 source_plan.join_condition, result_df, analyzer, expr_to_alias
-            )
+            ).fillna(value=False)
             sf_types = result_df.sf_types
             if "SEMI" in source_plan.join_type.sql:  # left semi
                 result_df = left[outer_join(left)]
@@ -1115,7 +1164,7 @@ def execute_mock_plan(
                 parameters_info={"source_plan.column_names": "True"},
                 raise_error=NotImplementedError,
             )
-        res_df = execute_mock_plan(source_plan.query)
+        res_df = execute_mock_plan(source_plan.query, expr_to_alias)
         return entity_registry.write_table(
             source_plan.table_name, res_df, source_plan.mode
         )
@@ -1125,7 +1174,7 @@ def execute_mock_plan(
             return entity_registry.read_table(entity_name)
         elif entity_registry.is_existing_view(entity_name):
             execution_plan = entity_registry.get_review(entity_name)
-            res_df = execute_mock_plan(execution_plan)
+            res_df = execute_mock_plan(execution_plan, expr_to_alias)
             return res_df
         else:
             obj_name_tuple = parse_table_name(entity_name)
@@ -1144,7 +1193,7 @@ def execute_mock_plan(
                 f"Object '{obj_database[1:-1]}.{obj_schema[1:-1]}.{obj_name[1:-1]}' does not exist or not authorized."
             )
     if isinstance(source_plan, Sample):
-        res_df = execute_mock_plan(source_plan.child)
+        res_df = execute_mock_plan(source_plan.child, expr_to_alias)
 
         if source_plan.row_count and (
             source_plan.row_count < 0 or source_plan.row_count > 100000
@@ -1185,7 +1234,7 @@ def execute_mock_plan(
             # Select rows to be updated based on condition
             condition = calculate_expression(
                 source_plan.condition, intermediate, analyzer, expr_to_alias
-            )
+            ).fillna(value=False)
 
             matched = target.apply(tuple, 1).isin(
                 intermediate[condition][target.columns].apply(tuple, 1)
@@ -1243,7 +1292,7 @@ def execute_mock_plan(
         if source_plan.condition:
             condition = calculate_expression(
                 source_plan.condition, intermediate, analyzer, expr_to_alias
-            )
+            ).fillna(value=False)
             intermediate = intermediate[condition]
             matched = target.apply(tuple, 1).isin(
                 intermediate[target.columns].apply(tuple, 1)
@@ -1292,7 +1341,7 @@ def execute_mock_plan(
                 if clause.condition:
                     condition = calculate_expression(
                         clause.condition, join_result, analyzer, expr_to_alias
-                    )
+                    ).fillna(value=False)
                     rows_to_update = join_result[condition]
                 else:
                     rows_to_update = join_result
@@ -1322,7 +1371,7 @@ def execute_mock_plan(
                 if clause.condition:
                     condition = calculate_expression(
                         clause.condition, join_result, analyzer, expr_to_alias
-                    )
+                    ).fillna(value=False)
                     intermediate = join_result[condition]
                 else:
                     intermediate = join_result
@@ -1354,7 +1403,7 @@ def execute_mock_plan(
                         unmatched_rows_in_source,
                         analyzer,
                         expr_to_alias,
-                    )
+                    ).fillna(value=False)
                     unmatched_rows_in_source = unmatched_rows_in_source[condition]
 
                 # filter out the unmatched rows that have been inserted in previous clauses
@@ -1433,7 +1482,7 @@ def execute_mock_plan(
 
         return [Row(*res)]
     elif isinstance(source_plan, Pivot):
-        child_rf = execute_mock_plan(source_plan.child)
+        child_rf = execute_mock_plan(source_plan.child, expr_to_alias)
 
         assert (
             len(source_plan.aggregates) == 1
@@ -1654,10 +1703,13 @@ def calculate_expression(
         )
         res = []
         for data in child_column:
-            try:
-                res.append(math.isnan(data))
-            except TypeError:
-                res.append(False)
+            if data is None:
+                res.append(None)
+            else:
+                try:
+                    res.append(math.isnan(data))
+                except TypeError:
+                    res.append(False)
         return ColumnEmulator(
             data=res, dtype=object, sf_type=ColumnType(BooleanType(), True)
         )
@@ -1894,7 +1946,7 @@ def calculate_expression(
                 break
             condition = calculate_expression(
                 case[0], input_data, analyzer, expr_to_alias
-            )
+            ).fillna(value=False)
             value = calculate_expression(case[1], input_data, analyzer, expr_to_alias)
 
             true_index = remaining[condition].index