+++ tests/integ/scala/test_dataframe_suite.py
@@ -1684,11 +1684,7 @@ def test_flatten_in_session(session):
     )
 
 
-@pytest.mark.skipif(
-    "config.getoption('local_testing_mode', default=False)",
-    reason="BUG: RecursionError: maximum recursion depth exceeded while calling a Python object",
-)
-def test_createDataFrame_with_given_schema(session):
+def test_createDataFrame_with_given_schema(session, local_testing_mode):
     schema = StructType(
         [
             StructField("string", StringType(84)),
@@ -1763,7 +1759,10 @@ def test_createDataFrame_with_given_schema(session):
             StructField("boolean", BooleanType()),
             StructField("binary", BinaryType()),
             StructField(
-                "timestamp", TimestampType(TimestampTimeZone.NTZ)
+                "timestamp",
+                TimestampType(TimestampTimeZone.NTZ)
+                if not local_testing_mode
+                else TimestampType(),
             ),  # depends on TIMESTAMP_TYPE_MAPPING
             StructField("timestamp_ntz", TimestampType(TimestampTimeZone.NTZ)),
             StructField("timestamp_ltz", TimestampType(TimestampTimeZone.LTZ)),
@@ -1791,11 +1790,7 @@ def test_createDataFrame_with_given_schema_time(session):
     assert df.collect() == data
 
 
-@pytest.mark.skipif(
-    "config.getoption('local_testing_mode', default=False)",
-    reason="BUG: assertion error timestamp type mismatch",
-)
-def test_createDataFrame_with_given_schema_timestamp(session):
+def test_createDataFrame_with_given_schema_timestamp(session, local_testing_mode):
     schema = StructType(
         [
             StructField("timestamp", TimestampType()),
@@ -1813,9 +1808,10 @@ def test_createDataFrame_with_given_schema_timestamp(session):
     ]
     df = session.create_dataframe(data, schema)
     schema_str = str(df.schema)
+
     assert (
         schema_str
-        == "StructType([StructField('TIMESTAMP', TimestampType(tz=ntz), nullable=True), "
+        == f"StructType([StructField('TIMESTAMP', TimestampType({'' if local_testing_mode else 'tz=ntz'}), nullable=True), "
         "StructField('TIMESTAMP_NTZ', TimestampType(tz=ntz), nullable=True), "
         "StructField('TIMESTAMP_LTZ', TimestampType(tz=ltz), nullable=True), "
         "StructField('TIMESTAMP_TZ', TimestampType(tz=tz), nullable=True)])"